requirements.txt
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.43.0

tantivy>=0.20.0

datasets>=2.18.0
huggingface-hub>=0.22.0

pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0
nltk>=3.8.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0

rouge-score>=0.1.2
bert-score>=0.3.13
evaluate>=0.4.0
scipy>=1.11.0
scikit-learn>=1.3.0

matplotlib>=3.7.0
seaborn>=0.13.0

biopython>=1.83
